make the grid faster - tell the neighbors that the cell exists, and then doing a second iteration to comb dead things
